Test on more versions of php
We don't need to increase the test matrix to include every possible scenario, but adding in more versions of php will help us catch unneeded errors.

This adds 6 new test scenarios, 2 of which are allowed to fail.

I chose to test 5.2 only against 4.0 since trunk is more likely to break against 5.2 than any other version of php and this will prevent that from being an issue for us.  Otherwise, we only test against trunk in the new versions.